test: consolidate reset button coverage
