--- tests/helpers/classicBattle/resetGameNextButton.test.js
@@ -1,35 +1,0 @@
-import { describe, it, expect, vi, beforeEach } from "vitest";
-
-describe("resetGame reattaches Next button handler", () => {
-  beforeEach(() => {
-    vi.resetModules();
-    document.body.innerHTML = "";
-  });
-
-  it("clones Next button and wires onNextButtonClick", async () => {
-    vi.mock("../../../src/helpers/classicBattle/timerService.js", () => ({
-      onNextButtonClick: vi.fn()
-    }));
-
-    // Original Next button before reset
-    const btn = document.createElement("button");
-    btn.id = "next-button";
-    btn.dataset.nextReady = "true"; // simulate previously ready state
-    document.body.appendChild(btn);
-
-    const { resetGame } = await import("../../../src/helpers/classicBattle/roundManager.js");
-    const timerSvc = await import("../../../src/helpers/classicBattle/timerService.js");
-
-    // Act: reset replaces the button and reattaches the click handler
-    resetGame();
-
-    const cloned = document.getElementById("next-button");
-    expect(cloned).not.toBe(btn); // ensure it was replaced
-    expect(cloned.disabled).toBe(true);
-    expect(cloned.dataset.nextReady).toBeUndefined();
-
-    // Clicking should invoke the reattached handler
-    cloned.dispatchEvent(new MouseEvent("click"));
-    expect(timerSvc.onNextButtonClick).toHaveBeenCalledTimes(1);
-  });
-});
